extends default EloquentFactory and FactoryBuilder
